requirements.txt
requests
boto3
pandas
sqlalchemy
psycopg2-binary
python-dotenv